app/backend/requirements.txt
azure-identity==1.14.0
azure-search-documents==11.6.0b4
quart==0.19.3
openai[datalib]==0.27.8
uvicorn[standard]==0.23.2
aiohttp==3.10.0